move ctx builder to its own mod
