--- crates/lune-std/src/context.rs
@@ -1,228 +1,0 @@
-use mlua::prelude::*;
-use std::{
-    borrow::Cow,
-    collections::HashMap,
-    path::{Path, PathBuf},
-};
-
-/**
-    Will only insert the item into the hashmap if the provided feature flag is enabled
-
-    # Example
-    ```
-    context_builder.with_alias("lune", |modules| {
-        insert_feature_only_module!(modules, "fs", LuneModuleCreator::LuaTable(lune_std_fs::module));
-
-        /*
-                turns into:
-
-        #[cfg(feature = "fs")]
-        modules.insert("fs", LuneModuleCreator::LuaTable(lune_std_fs::module));
-
-        */
-
-        Ok(())
-    })?;
-    ```
-*/
-#[macro_export]
-macro_rules! insert_feature_only_module {
-    ($modules:ident, $feature:literal, $module:expr) => {
-        #[cfg(feature = $feature)]
-        $modules.insert($feature, $module);
-    };
-}
-
-/**
-    Will insert the item into the hashmap
-
-    # Example
-    ```
-    context_builder.with_alias("lune", |modules| {
-        insert_module!(modules, "fs", LuneModuleCreator::LuaTable(lune_std_fs::module));
-
-        /*
-                turns into:
-
-        modules.insert("fs", LuneModuleCreator::LuaTable(lune_std_fs::module));
-
-        */
-
-        Ok(())
-    })?;
-    ```
-*/
-#[macro_export]
-macro_rules! insert_module {
-    ($modules:ident, $feature:literal, $module:expr) => {
-        $modules.insert($feature, $module);
-    };
-}
-
-/**
-    Use this enum to determine what type of lua object a module will return
-*/
-#[derive(Clone, Debug)]
-pub enum LuneModuleCreator {
-    LuaTable(fn(&Lua) -> LuaResult<LuaTable>),
-    LuaValue(fn(&Lua) -> LuaResult<LuaValue>),
-}
-
-impl<'lua> IntoLua<'lua> for LuneModuleCreator {
-    fn into_lua(self, lua: &'lua Lua) -> LuaResult<LuaValue<'lua>> {
-        match self {
-            LuneModuleCreator::LuaTable(creator) => creator(lua)?.into_lua(lua),
-            LuneModuleCreator::LuaValue(creator) => creator(lua),
-        }
-    }
-}
-
-#[derive(Default, Clone, Debug)]
-pub struct LuneModule {
-    pub children: HashMap<&'static str, LuneModuleCreator>,
-    pub alias: &'static str,
-}
-
-/**
-    This struct provides customizable information to globals at [`LuneStandardGlobal`](../enum.LuneStandardGlobal.html)
-
-    To create one, use [`GlobalsContextBuilder`](struct.GlobalsContextBuilder.html)
-*/
-#[derive(Default, Clone, Debug)]
-pub struct GlobalsContext {
-    pub modules: Vec<LuneModule>,
-    pub scripts: HashMap<PathBuf, Cow<'static, [u8]>>,
-}
-
-impl GlobalsContext {
-    #[must_use]
-    pub fn get_alias(&self, s: &str) -> Option<&LuneModule> {
-        self.modules.iter().find(|x| x.alias == s)
-    }
-
-    #[must_use]
-    pub fn get_script<T: Into<PathBuf>>(&self, abs_path: T) -> Option<&Cow<'static, [u8]>> {
-        let abs_path = abs_path.into();
-
-        self.scripts
-            .get(&abs_path)
-            .or(self.scripts.get(&abs_path.with_extension("luau")))
-            .or(self.scripts.get(&abs_path.with_extension("lua")))
-    }
-}
-
-/**
-    # Example
-    ```
-    // our module creator
-    let create_pixels = |lua: &Lua| -> LuaResult<LuaTable> {
-        ... // return a lua table
-    };
-
-    let builder = lune_std::context::GlobalsContextBuilder::new();
-
-    // lua: require("@<alias-name>/pixels")
-    builder.with_alias("<alias-name>", |modules| {
-        insert_module!(modules, "pixels", LuneModuleCreator::LuaTable(create_pixels));
-
-        Ok(())
-    })?;
-
-    lune_std::inject_globals(&lua, builder);
-    ```
-*/
-#[derive(Default)]
-pub struct GlobalsContextBuilder {
-    modules: Vec<LuneModule>,
-    scripts: HashMap<PathBuf, Cow<'static, [u8]>>,
-}
-
-impl GlobalsContextBuilder {
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /**
-        # Errors
-
-        Errors if the handler errors
-
-        # Example
-        ```
-        let create_pixels = |lua: &Lua| -> LuaResult<LuaTable> {
-            ... // return a lua table
-        };
-
-        builder.with_alias("<alias>", |modules| {
-            // There are multiple ways of inserting a module
-
-            // .1
-            modules.insert("pixels", LuneModuleCreator::LuaTable(create_pixels));
-
-            // .2
-            // does the exact same thing as .1
-            insert_module!(modules, "pixels", LuneModuleCreator::LuaTable(create_pixels));
-
-            // .3
-            // does the exact same thing as .1
-            // but only if a feature flag with the name of "pixels" is enabled
-            insert_feature_only_module!(modules, "pixels", LuneModuleCreator::LuaTable(create_pixels));
-
-            Ok(())
-        })?;
-        ```
-    */
-    pub fn with_alias(
-        &mut self,
-        name: &'static str,
-        handler: fn(&mut HashMap<&str, LuneModuleCreator>) -> LuaResult<()>,
-    ) -> LuaResult<()> {
-        let mut modules = HashMap::new();
-        handler(&mut modules)?;
-
-        let alias = LuneModule {
-            alias: name,
-            children: modules,
-        };
-
-        self.modules.push(alias);
-
-        Ok(())
-    }
-
-    /**
-        Define an absolute path for custom requires
-
-        # Example
-        ```
-        let script = "return 100";
-
-        globals_ctx_builder.with_script(
-            current_dir()
-                .unwrap()
-                .join("path/to/script"),
-            Cow::from(script.as_bytes()),
-        );
-
-        /*
-
-        -- lua code
-        local number = require("path/to/script")
-        print(number) -- 100
-
-         */
-        ```
-    */
-    pub fn with_script(&mut self, path: impl Into<PathBuf>, content: Cow<'static, [u8]>) {
-        self.scripts.insert(path.into(), content);
-    }
-
-    #[must_use]
-    pub fn build(self) -> GlobalsContext {
-        GlobalsContext {
-            modules: self.modules,
-            scripts: self.scripts,
-        }
-    }
-}
